requirements.txt
Django
djangorestframework
drf-yasg
pandas
pyspark
pyyaml
uritemplate